pyproject.toml
[project]
name = "mcp-chat-logger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.6.0",
    "pika>=1.3.0",
    "python-dotenv>=1.0.0",
]
